# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/users/orders": {
            "patch": {
                "tags": [
                    "Admin ChangeOrderStatus"
                ],
                "summary": "api for admin to change order status of user",
                "operationId": "ChangeOrderStatus",
                "parameters": [
                    {
                        "description": "inputs",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateOrderStatus"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order status updated successfully!",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Missing inputs",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Something went wrong!",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/carts/checkout": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User Cart"
                ],
                "summary": "api for user to checkout cart and proceed for payment",
                "operationId": "CheckoutCart",
                "responses": {
                    "200": {
                        "description": "Successfuly checked out",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Something went wrong! ",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/carts/checkout/razorpay": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User Cart"
                ],
                "summary": "api for create an razorpay order",
                "operationId": "RazorpayPage",
                "parameters": [
                    {
                        "description": "inputs",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RazorpayReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checkout successfull",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Something went wrong!",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/carts/checkout/razorpay/success": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User Cart"
                ],
                "summary": "api user for verify razorpay payment",
                "operationId": "RazorpayVerify",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment Method ID",
                        "name": "payment_method_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully payment completed and order approved",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to verify razor pay order!",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/carts/orders": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User orderHistory"
                ],
                "summary": "api for user to get all order history made",
                "operationId": "orderHistory",
                "responses": {
                    "200": {
                        "description": "Order history successful",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Missing user id",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Something went wrong! ",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/carts/place-order/cod": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User Cart"
                ],
                "summary": "api for user to place an order on cart with COD",
                "operationId": "PlaceCODOrder",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment option ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully order placed in COD",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Something went wrong! ",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.RazorpayReq": {
            "type": "object",
            "required": [
                "payment_method_id"
            ],
            "properties": {
                "payment_method_id": {
                    "type": "integer"
                }
            }
        },
...
        "request.UpdateOrderStatus": {
            "type": "object",
            "required": [
                "order_id",
                "status_id",
                "user_id"
            ],
            "properties": {
                "order_id": {
                    "type": "integer"
                },
                "status_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        }
    }
}